requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
APScheduler==3.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.1
Flask-Session==0.8.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
lxml==6.0.0
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
msgspec==0.19.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.0.2
openai==1.81.0
opencv-python-headless==4.12.0.88
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==27.0.1
redis==6.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.26.0
secure-smtplib==0.1.1
sib-api-v3-sdk==7.6.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
stack-data==0.6.3
stripe==12.1.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzlocal==5.3.1
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
yarg==0.1.9
zipp==3.21.0
